Добавил тесты к каждому классу
