+++ src/service/Managers.java
@@ -2,10 +2,7 @@ package service;
 
 public class Managers {
 
-    static InMemoryTaskManager inMemoryTaskManager;
-
     public static TaskManager getDefault() {
-        inMemoryTaskManager = new InMemoryTaskManager(new InMemoryHistoryManager());
-        return inMemoryTaskManager;
+        return new InMemoryTaskManager(new InMemoryHistoryManager());
     }
 }
